refactor: moving ColumnWrapper to types file
