+++ ui/src/lib/features/k8s/types.ts
@@ -7,3 +7,5 @@ export interface CommonRow {
     text: string
   }
 }
+
+export type ColumnWrapper<T> = [name: keyof T, styles?: string][]
